# ShriperApi/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Recipes": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "operationId": "Recipes_GetRecipes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recipes"
        ],
        "operationId": "Recipes_CreateRecipe",
        "requestBody": {
          "x-name": "createRecipeDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRecipeValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Recipes/{id}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "operationId": "Recipes_GetRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Recipe": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseModel"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "title",
              "ingredients",
              "instructions"
            ],
            "properties": {
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "title": {
                "type": "string",
                "maxLength": 100,
                "minLength": 0
              },
              "ingredients": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              },
              "instructions": {
                "type": "string",
                "maxLength": 1000,
                "minLength": 0
              },
              "vitalInstructions": {
                "type": "string",
                "maxLength": 200,
                "minLength": 0
              }
            }
          }
        ]
      },
      "Ingredient": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseModel"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "name",
              "amount"
            ],
            "properties": {
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "amount": {
                "type": "string",
                "minLength": 1
              }
            }
          }
        ]
      },
      "BaseModel": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateRecipeValidationErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateIngredientDtoValidationError"
            }
          },
          "instructions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vitalInstructions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateIngredientDtoValidationError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "amount": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateRecipeDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "ingredients",
          "instructions"
        ],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "ingredients": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/CreateIngredientDto"
            }
          },
          "instructions": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "vitalInstructions": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        }
      },
      "CreateIngredientDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "amount"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "amount": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
    }
  }
}